requirements.txt
Flask==1.1.1
requests==2.23.0
gunicorn==20.0.4
googletrans
google
bs4
pillow
